requirements.txt
# Context Quilt - Python Dependencies
# Generated for Python 3.11+

# ============================================
# Web Framework
# ============================================
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# ============================================
# LLM Provider SDKs
# ============================================
openai==1.3.5
anthropic==0.7.7
# google-generativeai==0.3.0  # Uncomment if using Gemini
# cohere==4.37  # Uncomment if using Cohere

# ============================================
# Database & Caching
# ============================================
psycopg2-binary==2.9.9
asyncpg==0.29.0
sqlalchemy==2.0.23
alembic==1.12.1
redis==5.0.1
hiredis==2.2.3  # Optional Redis performance boost

# ============================================
# Data Validation & Serialization
# ============================================
pydantic==2.5.0
pydantic-settings==2.1.0

# ============================================
# HTTP Client
# ============================================
httpx==0.25.2
aiohttp==3.9.1

# ============================================
# Authentication & Security
# ============================================
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# ============================================
# Monitoring & Observability
# ============================================
prometheus-client==0.19.0
structlog==23.2.0

# ============================================
# Utilities
# ============================================
python-dateutil==2.8.2
pytz==2023.3
tenacity==8.2.3  # Retry logic
cachetools==5.3.2

# ============================================
# Optional: Advanced Features
# ============================================
# tiktoken==0.5.2  # Token counting for OpenAI
# presidio-analyzer==2.2.33  # PII detection
# presidio-anonymizer==2.2.33
# sentence-transformers==2.2.2  # For semantic search

# ============================================
# Development Dependencies (move to requirements-dev.txt)
# ============================================
# pytest==7.4.3
# pytest-asyncio==0.21.1
# pytest-cov==4.1.0
# black==23.12.0
# flake8==6.1.0
# mypy==1.7.1
# isort==5.13.2
